# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Disciplr API
  version: 0.1.0
  description: API documentation for Disciplr backend
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/admin/users:
    get:
      summary: List all users (admin)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum:
              - USER
              - VERIFIER
              - ADMIN
        - name: status
          in: query
          schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
              - SUSPENDED
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: includeDeleted
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: List of users
  /api/admin/users/{id}/role:
    patch:
      summary: Update user role
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              def:
                type: object
                shape:
                  role:
                    def:
                      type: enum
                      entries:
                        USER: USER
                        VERIFIER: VERIFIER
                        ADMIN: ADMIN
                    type: enum
                    enum:
                      USER: USER
                      VERIFIER: VERIFIER
                      ADMIN: ADMIN
                    options:
                      - USER
                      - VERIFIER
                      - ADMIN
              type: object
      responses:
        "200":
          description: Role updated
        "400":
          description: Invalid role
        "404":
          description: User not found
  /api/admin/users/{id}/status:
    patch:
      summary: Update user status
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              def:
                type: object
                shape:
                  status:
                    def:
                      type: enum
                      entries:
                        ACTIVE: ACTIVE
                        INACTIVE: INACTIVE
                        SUSPENDED: SUSPENDED
                    type: enum
                    enum:
                      ACTIVE: ACTIVE
                      INACTIVE: INACTIVE
                      SUSPENDED: SUSPENDED
                    options:
                      - ACTIVE
                      - INACTIVE
                      - SUSPENDED
              type: object
      responses:
        "200":
          description: Status updated
        "404":
          description: User not found
  /api/admin/users/{id}:
    delete:
      summary: Delete user (soft or hard)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: hard
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: User deleted
        "400":
          description: Cannot delete own account
        "404":
          description: User not found
  /api/admin/users/{id}/restore:
    post:
      summary: Restore a soft-deleted user
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User restored
        "400":
          description: User not deleted
        "404":
          description: User not found
  /api/admin/users/{userId}/revoke-sessions:
    post:
      summary: Revoke all user sessions
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sessions revoked
  /api/admin/audit-logs:
    get:
      summary: List audit logs
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: actor_user_id
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: target_type
          in: query
          schema:
            type: string
        - name: target_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  audit_logs:
                    type: array
                    items: {}
                  count:
                    type: number
                  total:
                    type: number
                  limit:
                    type: number
                  offset:
                    type: number
                  has_more:
                    type: boolean
                required:
                  - audit_logs
                  - count
                  - total
                  - offset
                  - has_more
  /api/admin/audit-logs/{id}:
    get:
      summary: Get audit log by ID
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audit log details
        "404":
          description: Not found
  /api/admin/overrides/vaults/{id}/cancel:
    post:
      summary: Admin override to cancel a vault
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              def:
                type: object
                shape:
                  reasonCode:
                    def:
                      type: enum
                      entries:
                        USER_REQUEST: USER_REQUEST
                        FRAUD_DETECTED: FRAUD_DETECTED
                        SYSTEM_ERROR: SYSTEM_ERROR
                        POLICY_VIOLATION: POLICY_VIOLATION
                        EMERGENCY_ADMIN_ACTION: EMERGENCY_ADMIN_ACTION
                        COMPLIANCE_REQUIREMENT: COMPLIANCE_REQUIREMENT
                        TESTING_CLEANUP: TESTING_CLEANUP
                    type: enum
                    enum:
                      USER_REQUEST: USER_REQUEST
                      FRAUD_DETECTED: FRAUD_DETECTED
                      SYSTEM_ERROR: SYSTEM_ERROR
                      POLICY_VIOLATION: POLICY_VIOLATION
                      EMERGENCY_ADMIN_ACTION: EMERGENCY_ADMIN_ACTION
                      COMPLIANCE_REQUIREMENT: COMPLIANCE_REQUIREMENT
                      TESTING_CLEANUP: TESTING_CLEANUP
                    options:
                      - USER_REQUEST
                      - FRAUD_DETECTED
                      - SYSTEM_ERROR
                      - POLICY_VIOLATION
                      - EMERGENCY_ADMIN_ACTION
                      - COMPLIANCE_REQUIREMENT
                      - TESTING_CLEANUP
                  reason:
                    def:
                      type: optional
                      innerType:
                        def:
                          type: string
                        type: string
                        format: null
                        minLength: null
                        maxLength: null
                    type: optional
                  idempotencyKey:
                    def:
                      type: optional
                      innerType:
                        def:
                          type: string
                        type: string
                        format: null
                        minLength: null
                        maxLength: null
                    type: optional
                  details:
                    def:
                      type: optional
                      innerType:
                        def:
                          type: string
                        type: string
                        format: null
                        minLength: null
                        maxLength: null
                    type: optional
              type: object
      responses:
        "200":
          description: Vault cancelled
        "400":
          description: Invalid reason code
        "404":
          description: Vault not found
        "409":
          description: Vault already cancelled or not cancellable
...